==> 1_data-clean.R <==
#Clear existing data and graphics
# rm(list=ls())
# graphics.off()


#load helper file
source("0_pachages-function.R")


###########
#
# Load redcap data, scrub text, set labels and factors
#
##########


#Read Data
data=read.csv('redcap/AutoimmunityAndHealt_DATA_2022-07-01_0751.csv')


#scrub text
data <- data %>%
  mutate_all(tolower) %>%
  mutate_all(~str_replace_all(., "\\?", "")) %>%
  mutate_all(~str_replace_all(., "\\.", "")) %>%
  mutate_all(~str_replace_all(., "\\'", "")) %>%
  mutate_all(str_trim) %>%
  mutate_all(str_squish)


#Setting Labels
label(data$record_id)="Record ID"
label(data$redcap_survey_identifier)="Survey Identifier"
label(data$intro_timestamp)="Survey Timestamp"
label(data$emailid)="Please provide your email for follow-up and to login.  Note, we will only use this email to request formal consent or follow-up surveys. "
label(data$residence)="Is Australia your primary residence?"
label(data$sex_diff_bool)="Are you interested in participating in a study on sex differences in autoimmune disorders? "
label(data$intro_complete)="Complete?"
label(data$consent_form_timestamp)="Survey Timestamp"
label(data$consentb)="I consent to the storage and use of survey responses taken from me for use in:"
label(data$consentb_2)="I wish to be informed of any significant results generated by this research:"
label(data$consentb_3)="I understand that I will be given a signed copy of this document to keep."
label(data$consent_form_complete)="Complete?"
label(data$demographics_timestamp)="Survey Timestamp"
label(data$state)="State or territory"
label(data$postcode)="Postcode"
label(data$year_of_birth)="Year of birth"
label(data$rel_status)="Which of the following best describes your current relationship status?"
label(data$home)="Which of the following best describes your living situation? "
label(data$education)="What is the highest level of schooling you have completed or the highest degree you have received?"
label(data$employment_status)="Which of the following categories best describes your employment status?"
label(data$gender)="As what gender do you identify?"
label(data$ethnicity)="Which categories describe you? "
label(data$lgbtq)="Do you identify as LGBTQ+ (lesbian, gay, bisexual, transgender, queer, plus)?"
label(data$mental_id)="Do you possess a physical or mental impairment that limits life activities?"
label(data$alcohol)="Do you consume alcohol?"
label(data$smoke)="Do you smoke?"
label(data$diseasestat)="Do you have a chronic illness or underlying medical condition?"
label(data$demographics_complete)="Complete?"
label(data$disease_and_diagnosis_timestamp)="Survey Timestamp"
label(data$self)="Self-diagnosed"
label(data$formal)="Formally diagnosed"
label(data$autoimmune_id___1)="What autoimmune disease (or diseases) have you been formally diagnosed with? Please select all that apply. If your condition is not in the list, select other and specify.  (choice=Ankylosing spondylitis)"
label(data$autoimmune_id___2)="What autoimmune disease (or diseases) have you been formally diagnosed with? Please select all that apply. If your condition is not in the list, select other and specify.  (choice=Autoimmune haemolytic anaemia)"
label(data$autoimmune_id___3)="What autoimmune disease (or diseases) have you been formally diagnosed with? Please select all that apply. If your condition is not in the list, select other and specify.  (choice=Autoimmune vitiligo)"
label(data$autoimmune_id___4)="What autoimmune disease (or diseases) have you been formally diagnosed with? Please select all that apply. If your condition is not in the list, select other and specify.  (choice=Bullous pemphigoid)"
label(data$autoimmune_id___5)="What autoimmune disease (or diseases) have you been formally diagnosed with? Please select all that apply. If your condition is not in the list, select other and specify.  (choice=Chronic inflammatory demyelinating neuropathy)"
label(data$autoimmune_id___6)="What autoimmune disease (or diseases) have you been formally diagnosed with? Please select all that apply. If your condition is not in the list, select other and specify.  (choice=Coeliac disease)"
label(data$autoimmune_id___7)="What autoimmune disease (or diseases) have you been formally diagnosed with? Please select all that apply. If your condition is not in the list, select other and specify.  (choice=CREST syndrome (limited scleroderma))"
label(data$autoimmune_id___8)="What autoimmune disease (or diseases) have you been formally diagnosed with? Please select all that apply. If your condition is not in the list, select other and specify.  (choice=Crohns disease)"
label(data$autoimmune_id___9)="What autoimmune disease (or diseases) have you been formally diagnosed with? Please select all that apply. If your condition is not in the list, select other and specify.  (choice=Dermatomyositis)"
label(data$autoimmune_id___10)="What autoimmune disease (or diseases) have you been formally diagnosed with? Please select all that apply. If your condition is not in the list, select other and specify.  (choice=Essential mixed cryoglobulinemia)"
label(data$autoimmune_id___11)="What autoimmune disease (or diseases) have you been formally diagnosed with? Please select all that apply. If your condition is not in the list, select other and specify.  (choice=Evans syndrome)"
label(data$autoimmune_id___12)="What autoimmune disease (or diseases) have you been formally diagnosed with? Please select all that apply. If your condition is not in the list, select other and specify.  (choice=GBM disease (Goodpastures syndrome))"
label(data$autoimmune_id___13)="What autoimmune disease (or diseases) have you been formally diagnosed with? Please select all that apply. If your condition is not in the list, select other and specify.  (choice=Giant cell arteritis)"
label(data$autoimmune_id___14)="What autoimmune disease (or diseases) have you been formally diagnosed with? Please select all that apply. If your condition is not in the list, select other and specify.  (choice=Graves disease)"
label(data$autoimmune_id___15)="What autoimmune disease (or diseases) have you been formally diagnosed with? Please select all that apply. If your condition is not in the list, select other and specify.  (choice=Guillain Barré syndrome)"
label(data$autoimmune_id___16)="What autoimmune disease (or diseases) have you been formally diagnosed with? Please select all that apply. If your condition is not in the list, select other and specify.  (choice=Infective endocarditis associated vasculitis)"
label(data$autoimmune_id___17)="What autoimmune disease (or diseases) have you been formally diagnosed with? Please select all that apply. If your condition is not in the list, select other and specify.  (choice=Hashimotos disease (autoimmune hypothyroidism))"
label(data$autoimmune_id___18)="What autoimmune disease (or diseases) have you been formally diagnosed with? Please select all that apply. If your condition is not in the list, select other and specify.  (choice=Henoch Schonlein purpura)"
label(data$autoimmune_id___19)="What autoimmune disease (or diseases) have you been formally diagnosed with? Please select all that apply. If your condition is not in the list, select other and specify.  (choice=IgA nephropathy)"
label(data$autoimmune_id___20)="What autoimmune disease (or diseases) have you been formally diagnosed with? Please select all that apply. If your condition is not in the list, select other and specify.  (choice=IgG4 related disease)"
label(data$autoimmune_id___21)="What autoimmune disease (or diseases) have you been formally diagnosed with? Please select all that apply. If your condition is not in the list, select other and specify.  (choice=Immune thrombocytopenia)"
label(data$autoimmune_id___22)="What autoimmune disease (or diseases) have you been formally diagnosed with? Please select all that apply. If your condition is not in the list, select other and specify.  (choice=Lupus (systemic lupus erythematosus))"
label(data$autoimmune_id___23)="What autoimmune disease (or diseases) have you been formally diagnosed with? Please select all that apply. If your condition is not in the list, select other and specify.  (choice=Membranous nephropathy)"
label(data$autoimmune_id___24)="What autoimmune disease (or diseases) have you been formally diagnosed with? Please select all that apply. If your condition is not in the list, select other and specify.  (choice=Microscopic polyangiitis)"
label(data$autoimmune_id___25)="What autoimmune disease (or diseases) have you been formally diagnosed with? Please select all that apply. If your condition is not in the list, select other and specify.  (choice=Motor neurone disease)"
label(data$autoimmune_id___26)="What autoimmune disease (or diseases) have you been formally diagnosed with? Please select all that apply. If your condition is not in the list, select other and specify.  (choice=Multiple sclerosis)"
label(data$autoimmune_id___27)="What autoimmune disease (or diseases) have you been formally diagnosed with? Please select all that apply. If your condition is not in the list, select other and specify.  (choice=Myasthenia gravis)"
label(data$autoimmune_id___28)="What autoimmune disease (or diseases) have you been formally diagnosed with? Please select all that apply. If your condition is not in the list, select other and specify.  (choice=Neuromyelitis optica (Devics disease))"
label(data$autoimmune_id___29)="What autoimmune disease (or diseases) have you been formally diagnosed with? Please select all that apply. If your condition is not in the list, select other and specify.  (choice=Pemphigus vulgaris)"
label(data$autoimmune_id___30)="What autoimmune disease (or diseases) have you been formally diagnosed with? Please select all that apply. If your condition is not in the list, select other and specify.  (choice=Pernicious anaemia)"
label(data$autoimmune_id___31)="What autoimmune disease (or diseases) have you been formally diagnosed with? Please select all that apply. If your condition is not in the list, select other and specify.  (choice=Polymyositis)"
label(data$autoimmune_id___32)="What autoimmune disease (or diseases) have you been formally diagnosed with? Please select all that apply. If your condition is not in the list, select other and specify.  (choice=Primary biliary cirrhosis)"
label(data$autoimmune_id___33)="What autoimmune disease (or diseases) have you been formally diagnosed with? Please select all that apply. If your condition is not in the list, select other and specify.  (choice=Primary antiphospholipid syndrome)"
label(data$autoimmune_id___34)="What autoimmune disease (or diseases) have you been formally diagnosed with? Please select all that apply. If your condition is not in the list, select other and specify.  (choice=Psoriatic arthritis)"
label(data$autoimmune_id___35)="What autoimmune disease (or diseases) have you been formally diagnosed with? Please select all that apply. If your condition is not in the list, select other and specify.  (choice=Rheumatoid arthritis)"
label(data$autoimmune_id___36)="What autoimmune disease (or diseases) have you been formally diagnosed with? Please select all that apply. If your condition is not in the list, select other and specify.  (choice=Sarcoidosis)"
label(data$autoimmune_id___37)="What autoimmune disease (or diseases) have you been formally diagnosed with? Please select all that apply. If your condition is not in the list, select other and specify.  (choice=Scleroderma (systemic sclerosis))"
label(data$autoimmune_id___38)="What autoimmune disease (or diseases) have you been formally diagnosed with? Please select all that apply. If your condition is not in the list, select other and specify.  (choice=Sjögrens syndrome)"
label(data$autoimmune_id___39)="What autoimmune disease (or diseases) have you been formally diagnosed with? Please select all that apply. If your condition is not in the list, select other and specify.  (choice=Susacs syndrome)"
label(data$autoimmune_id___40)="What autoimmune disease (or diseases) have you been formally diagnosed with? Please select all that apply. If your condition is not in the list, select other and specify.  (choice=Sporadic Inclusion Body Myositis)"
label(data$autoimmune_id___41)="What autoimmune disease (or diseases) have you been formally diagnosed with? Please select all that apply. If your condition is not in the list, select other and specify.  (choice=Type 1 diabetes)"
label(data$autoimmune_id___42)="What autoimmune disease (or diseases) have you been formally diagnosed with? Please select all that apply. If your condition is not in the list, select other and specify.  (choice=Ulcerative colitis)"
label(data$autoimmune_id___43)="What autoimmune disease (or diseases) have you been formally diagnosed with? Please select all that apply. If your condition is not in the list, select other and specify.  (choice=Wegeners granulomatosis)"
label(data$autoimmune_id___44)="What autoimmune disease (or diseases) have you been formally diagnosed with? Please select all that apply. If your condition is not in the list, select other and specify.  (choice=None)"
label(data$autoimmune_id___45)="What autoimmune disease (or diseases) have you been formally diagnosed with? Please select all that apply. If your condition is not in the list, select other and specify.  (choice=Other)"
label(data$autoimmune_id_other)="If other, please specify: "
label(data$diagnosis_date)="If you have been formally diagnosed, please provide diagnosis details: approximate date.  "
label(data$diagnosis_type)="If you have been formally diagnosed, please provide diagnosis details: type of specialist/medical professional. "
label(data$diagnosis_details)="If you have been formally diagnosed, please provide diagnosis details youd like to share. "
label(data$system)="If you have been diagnosed, or are currently seeking a diagnosis, please specify whether youre pursuing your medical care through the public or private system."
label(data$diagnosis_length)="If youve been formally diagnosed, how long did it/has it been taking between onset of symptoms and diagnosis? If more than one, please select the first diagnosis. "
label(data$viral1)="Viral infection?"
label(data$preg1)="During a pregnancy?"
label(data$preg2)="After a pregnancy?"
label(data$meno1)="After menopause?"
label(data$ill1)="An illness?"
label(data$otherlife)="Other?"
label(data$otherchronic_2)="If other, please specify:"
label(data$misdiag)="If you were formally diagnosed, were you misdiagnosed at any stage?"
label(data$chronic___1)="Do you have, or have you ever had a chronic or severe illness (including genetic diseases or disorders)?  (choice=Asthma)"
label(data$chronic___2)="Do you have, or have you ever had a chronic or severe illness (including genetic diseases or disorders)?  (choice=Diabetes (type 2))"
label(data$chronic___3)="Do you have, or have you ever had a chronic or severe illness (including genetic diseases or disorders)?  (choice=Cancer)"
label(data$chronic___4)="Do you have, or have you ever had a chronic or severe illness (including genetic diseases or disorders)?  (choice=Mental illness)"
label(data$chronic___5)="Do you have, or have you ever had a chronic or severe illness (including genetic diseases or disorders)?  (choice=Heart disease)"
label(data$chronic___7)="Do you have, or have you ever had a chronic or severe illness (including genetic diseases or disorders)?  (choice=Other)"
label(data$chronic___8)="Do you have, or have you ever had a chronic or severe illness (including genetic diseases or disorders)?  (choice=None)"
label(data$otherchronic)="If other, please specify:"
label(data$disease_and_diagnosis_complete)="Complete?"
label(data$symptom_management_timestamp)="Survey Timestamp"
label(data$fatigue)="Fatigue"
label(data$joint)="Joint pain and swelling"
label(data$skin)="Skin problems"
label(data$stomach)="Abdominal pain or digestive issues"
label(data$fever)="Recurring fever"
label(data$glands)="Swollen glands"
label(data$chills)="Chills"
label(data$memory)="Memory loss"
label(data$conc)="Concentration loss"
label(data$hair)="Hair loss"
label(data$skin2)="Peculiar skin sensation"
label(data$weak)="General weakness"
label(data$pain)="Muscle aches and pain"
label(data$bruise)="Easy bruising"
label(data$treatments___1)="How do you treat your symptoms? (choice=Pain relief)"
label(data$treatments___2)="How do you treat your symptoms? (choice=Immunosuppressants)"
label(data$treatments___3)="How do you treat your symptoms? (choice=Hormone replacements and/or injections)"
label(data$treatments___5)="How do you treat your symptoms? (choice=None)"
label(data$treatments___4)="How do you treat your symptoms? (choice=Other)"
label(data$symptoms_other)="If other, please specify:"
label(data$flares___1)="Flares are the sudden and severe onset of symptoms. What triggers your symptoms/flares? (choice=Stress)"
label(data$flares___2)="Flares are the sudden and severe onset of symptoms. What triggers your symptoms/flares? (choice=Sun)"
label(data$flares___3)="Flares are the sudden and severe onset of symptoms. What triggers your symptoms/flares? (choice=Drugs)"
label(data$flares___4)="Flares are the sudden and severe onset of symptoms. What triggers your symptoms/flares? (choice=Unknown)"
label(data$flares___6)="Flares are the sudden and severe onset of symptoms. What triggers your symptoms/flares? (choice=None)"
label(data$flares___5)="Flares are the sudden and severe onset of symptoms. What triggers your symptoms/flares? (choice=Other)"
label(data$flares_other)="If other, please specify:"
label(data$drug___1)="Are you on any medication to treat your symptoms?  (choice=Over the counter (OTC) such as aspirin or ibuprofen)"
label(data$drug___2)="Are you on any medication to treat your symptoms?  (choice=Prescription)"
label(data$drug___3)="Are you on any medication to treat your symptoms?  (choice=Non-prescribed prescription medication)"
label(data$drug___4)="Are you on any medication to treat your symptoms?  (choice=None)"
label(data$drug___5)="Are you on any medication to treat your symptoms?  (choice=Other)"
label(data$treatments_other_other)="If other, please specify:"
label(data$med_other)="Are you on medication for other reasons?"
label(data$drug_other)="Do you take any recreational/illicit drugs to treat your symptoms?"
label(data$drug_other_other)="If yes, please specify:"
label(data$hormone___1)="Are you on any hormone treatment? (choice=Contraceptive pill)"
label(data$hormone___3)="Are you on any hormone treatment? (choice=Testosterone therapy)"
label(data$hormone___2)="Are you on any hormone treatment? (choice=Hormone replacement therapy (HRT))"
label(data$hormone___4)="Are you on any hormone treatment? (choice=Thyroid hormone replacement)"
label(data$hormone___5)="Are you on any hormone treatment? (choice=Insulin)"
label(data$hormone___6)="Are you on any hormone treatment? (choice=None)"
label(data$hormone___7)="Are you on any hormone treatment? (choice=Other)"
label(data$hormone_other)="If other, please specify:"
label(data$diet___1)="Do you use a special diet to manage your symptoms? (choice=Nutritional supplements)"
label(data$diet___2)="Do you use a special diet to manage your symptoms? (choice=Gluten free)"
label(data$diet___3)="Do you use a special diet to manage your symptoms? (choice=FODMAP)"
label(data$diet___4)="Do you use a special diet to manage your symptoms? (choice=Vegan)"
label(data$diet___5)="Do you use a special diet to manage your symptoms? (choice=Sugar free or low carbohydrate)"
label(data$diet___6)="Do you use a special diet to manage your symptoms? (choice=Keto)"
label(data$diet___7)="Do you use a special diet to manage your symptoms? (choice=None)"
label(data$diet___8)="Do you use a special diet to manage your symptoms? (choice=Other specialty diet)"
label(data$diet_other)="If other, please specify:"
label(data$physical___1)="Do you use physical therapy to manage your symptoms? (choice=Physiotherapy)"
label(data$physical___2)="Do you use physical therapy to manage your symptoms? (choice=Exercise)"
label(data$physical___3)="Do you use physical therapy to manage your symptoms? (choice=Massage therapy)"
label(data$physical___4)="Do you use physical therapy to manage your symptoms? (choice=None)"
label(data$physical___5)="Do you use physical therapy to manage your symptoms? (choice=Other)"
label(data$physical_other)="If other, please specify:"
label(data$specialist___1)="Are you being treated or were ever treated by a specialist? Select all that apply.  (choice=Nephrologist (kidney problems))"
label(data$specialist___2)="Are you being treated or were ever treated by a specialist? Select all that apply.  (choice=Rheumatologist (arthritis and other rheumatic diseases))"
label(data$specialist___3)="Are you being treated or were ever treated by a specialist? Select all that apply.  (choice=Endocrinologist (gland and hormone problems))"
label(data$specialist___4)="Are you being treated or were ever treated by a specialist? Select all that apply.  (choice=Neurologist (nerve problems))"
label(data$specialist___5)="Are you being treated or were ever treated by a specialist? Select all that apply.  (choice=Hematologist (diseases that affect blood, such as some forms of anemia))"
label(data$specialist___6)="Are you being treated or were ever treated by a specialist? Select all that apply.  (choice=Gastroenterologist (digestive system, such as inflammatory bowel disease))"
label(data$specialist___7)="Are you being treated or were ever treated by a specialist? Select all that apply.  (choice=Dermatologist (skin, hair, and nails))"
label(data$specialist___8)="Are you being treated or were ever treated by a specialist? Select all that apply.  (choice=Immunologist (immune system related))"
label(data$specialist___9)="Are you being treated or were ever treated by a specialist? Select all that apply.  (choice=None)"
label(data$specialist___10)="Are you being treated or were ever treated by a specialist? Select all that apply.  (choice=Other)"
label(data$specialist_other)="If other, please specify:"
label(data$therapy___1)="Are you being treated or were ever treated by a therapist? Select all that apply.  (choice=Physical therapist (stiffness, weakness, and restricted body movement))"
label(data$therapy___2)="Are you being treated or were ever treated by a therapist? Select all that apply.  (choice=Occupational therapist (activities of daily living))"
label(data$therapy___3)="Are you being treated or were ever treated by a therapist? Select all that apply.  (choice=Speech therapist (speech problems))"
label(data$therapy___4)="Are you being treated or were ever treated by a therapist? Select all that apply.  (choice=Audiologist (hearing problems, including inner ear damage))"
label(data$therapy___5)="Are you being treated or were ever treated by a therapist? Select all that apply.  (choice=Vocational therapist (job training))"
label(data$therapy___6)="Are you being treated or were ever treated by a therapist? Select all that apply.  (choice=Counselor for emotional support (or coping with your illness))"
label(data$therapy___7)="Are you being treated or were ever treated by a therapist? Select all that apply.  (choice=None)"
label(data$therapy___8)="Are you being treated or were ever treated by a therapist? Select all that apply.  (choice=Other)"
label(data$therapy_other)="If other, please specify:"
label(data$treatments_other)="Other treatments youd like to disclose not included: "
label(data$symptom_management_complete)="Complete?"
label(data$mental_health_timestamp)="Survey Timestamp"
label(data$selfmental)="Self-diagnosed"
label(data$formalmental)="Formally diagnosed"
label(data$mental_diagn___1)="What mental health illness class (or illnesses) have you been formally diagnosed with? (choice=Anxiety disorders (e.g., panic disorder, obsessive-compulsive disorder, and phobias))"
label(data$mental_diagn___2)="What mental health illness class (or illnesses) have you been formally diagnosed with? (choice=Mood disorders (e.g., depression, bipolar disorder))"
label(data$mental_diagn___3)="What mental health illness class (or illnesses) have you been formally diagnosed with? (choice=Eating disorders (e.g., anorexia nervosa))"
label(data$mental_diagn___4)="What mental health illness class (or illnesses) have you been formally diagnosed with? (choice=Personality disorders (e.g., paranoia, schizophrenia))"
label(data$mental_diagn___5)="What mental health illness class (or illnesses) have you been formally diagnosed with? (choice=Post-traumatic stress disorder (PTSD))"
label(data$mental_diagn___6)="What mental health illness class (or illnesses) have you been formally diagnosed with? (choice=None)"
label(data$mental_diagn___7)="What mental health illness class (or illnesses) have you been formally diagnosed with? (choice=Other)"
label(data$mental_diagn_other)="If other, please specify: "
label(data$ptsd___1)="CW *** Post-traumatic stress disorder (PTSD) is a particular set of reactions that can develop in people who have been through a traumatic event which threatened their life or safety, or that of others around them.  Have you experienced or seen something that involved death, injury, torture or abuse and felt very scared or helpless? (choice=Car or other serious accident)"
label(data$ptsd___2)="CW *** Post-traumatic stress disorder (PTSD) is a particular set of reactions that can develop in people who have been through a traumatic event which threatened their life or safety, or that of others around them.  Have you experienced or seen something that involved death, injury, torture or abuse and felt very scared or helpless? (choice=Physical or sexual assault)"
label(data$ptsd___3)="CW *** Post-traumatic stress disorder (PTSD) is a particular set of reactions that can develop in people who have been through a traumatic event which threatened their life or safety, or that of others around them.  Have you experienced or seen something that involved death, injury, torture or abuse and felt very scared or helpless? (choice=War or torture (e.g., combat or refugee))"
label(data$ptsd___4)="CW *** Post-traumatic stress disorder (PTSD) is a particular set of reactions that can develop in people who have been through a traumatic event which threatened their life or safety, or that of others around them.  Have you experienced or seen something that involved death, injury, torture or abuse and felt very scared or helpless? (choice=Natural disasters (e.g., bushfires or floods))"
label(data$ptsd___5)="CW *** Post-traumatic stress disorder (PTSD) is a particular set of reactions that can develop in people who have been through a traumatic event which threatened their life or safety, or that of others around them.  Have you experienced or seen something that involved death, injury, torture or abuse and felt very scared or helpless? (choice=Medical trauma (e.g., first responder, medic, serious illness or procedures))"
label(data$ptsd___7)="CW *** Post-traumatic stress disorder (PTSD) is a particular set of reactions that can develop in people who have been through a traumatic event which threatened their life or safety, or that of others around them.  Have you experienced or seen something that involved death, injury, torture or abuse and felt very scared or helpless? (choice=None)"
label(data$ptsd___6)="CW *** Post-traumatic stress disorder (PTSD) is a particular set of reactions that can develop in people who have been through a traumatic event which threatened their life or safety, or that of others around them.  Have you experienced or seen something that involved death, injury, torture or abuse and felt very scared or helpless? (choice=Other)"
label(data$ptsd_other)="If other, please specify: "
label(data$ptsd_freq)="CW ***  If you suffer from PTSD, how often do you experience negative effects from the condition on your everyday life?"
label(data$ptsd_freq_other)="If other, please specify: "
label(data$selffamment)="Self-diagnosed"
label(data$formalfamment)="Formally diagnosed"
label(data$mental_diagn_fam___1)="What mental health illness class (or illnesses) has your family member or partner been formally diagnosed with? (choice=Anxiety disorders (e.g., panic disorder, obsessive-compulsive disorder, and phobias))"
label(data$mental_diagn_fam___2)="What mental health illness class (or illnesses) has your family member or partner been formally diagnosed with? (choice=Mood disorders (e.g., depression, bipolar disorder))"
label(data$mental_diagn_fam___3)="What mental health illness class (or illnesses) has your family member or partner been formally diagnosed with? (choice=Eating disorders (e.g., anorexia nervosa))"
label(data$mental_diagn_fam___4)="What mental health illness class (or illnesses) has your family member or partner been formally diagnosed with? (choice=Personality disorders (e.g., paranoia, schizophrenia))"
label(data$mental_diagn_fam___5)="What mental health illness class (or illnesses) has your family member or partner been formally diagnosed with? (choice=Post-traumatic stress disorder (PTSD))"
label(data$mental_diagn_fam___6)="What mental health illness class (or illnesses) has your family member or partner been formally diagnosed with? (choice=None)"
label(data$mental_diagn_fam___7)="What mental health illness class (or illnesses) has your family member or partner been formally diagnosed with? (choice=Other)"
label(data$mental_diagn_fam_other)="If other, please specify:"
label(data$ptsd_fam___1)="CW ***  Have one or more family members been diagnosed with PTSD or experienced or seen something that involved death, injury, torture or abuse and felt very scared or helpless? (choice=Diagnosed with PTSD)"
label(data$ptsd_fam___2)="CW ***  Have one or more family members been diagnosed with PTSD or experienced or seen something that involved death, injury, torture or abuse and felt very scared or helpless? (choice=Car or other serious accident)"
label(data$ptsd_fam___3)="CW ***  Have one or more family members been diagnosed with PTSD or experienced or seen something that involved death, injury, torture or abuse and felt very scared or helpless? (choice=Physical or sexual assault)"
label(data$ptsd_fam___4)="CW ***  Have one or more family members been diagnosed with PTSD or experienced or seen something that involved death, injury, torture or abuse and felt very scared or helpless? (choice=War or torture (e.g., combat or refugee))"
label(data$ptsd_fam___5)="CW ***  Have one or more family members been diagnosed with PTSD or experienced or seen something that involved death, injury, torture or abuse and felt very scared or helpless? (choice=Natural disasters (e.g., bushfires or floods))"
label(data$ptsd_fam___6)="CW ***  Have one or more family members been diagnosed with PTSD or experienced or seen something that involved death, injury, torture or abuse and felt very scared or helpless? (choice=Medical trauma (e.g., first responder, medic, serious illness or procedures))"
label(data$ptsd_fam___8)="CW ***  Have one or more family members been diagnosed with PTSD or experienced or seen something that involved death, injury, torture or abuse and felt very scared or helpless? (choice=None)"
label(data$ptsd_fam___7)="CW ***  Have one or more family members been diagnosed with PTSD or experienced or seen something that involved death, injury, torture or abuse and felt very scared or helpless? (choice=Other)"
label(data$ptsd_fam_other)="If other, please specify:"
label(data$mental_health_complete)="Complete?"
label(data$neuroqol_bank_v10_fatigue_timestamp)="Survey Timestamp"
label(data$neuroqol_bank_v10_fatigueol_tscore)="T-score"
label(data$neuroqol_bank_v10_fatigueol_std_error)="Standard Error"
label(data$neuroqol_nqftg13)="In the past 7 days I felt exhausted"
label(data$neuroqol_nqftg13_tscore)="T-score (Cumulative)"
label(data$neuroqol_nqftg13_stderror)="Standard Error (Cumulative)"
label(data$neuroqol_nqftg13_qposition)="Question Position"
label(data$neuroqol_nqftg11)="In the past 7 days I felt that I had no energy"
label(data$neuroqol_nqftg11_tscore)="T-score (Cumulative)"
label(data$neuroqol_nqftg11_stderror)="Standard Error (Cumulative)"
label(data$neuroqol_nqftg11_qposition)="Question Position"
label(data$neuroqol_nqftg15)="In the past 7 days I felt  fatigued"
label(data$neuroqol_nqftg15_tscore)="T-score (Cumulative)"
label(data$neuroqol_nqftg15_stderror)="Standard Error (Cumulative)"
label(data$neuroqol_nqftg15_qposition)="Question Position"
label(data$neuroqol_nqftg06)="In the past 7 days I was too tired to do my household chores"
label(data$neuroqol_nqftg06_tscore)="T-score (Cumulative)"
label(data$neuroqol_nqftg06_stderror)="Standard Error (Cumulative)"
label(data$neuroqol_nqftg06_qposition)="Question Position"
label(data$neuroqol_nqftg07)="In the past 7 days I was too tired to leave the house"
label(data$neuroqol_nqftg07_tscore)="T-score (Cumulative)"
label(data$neuroqol_nqftg07_stderror)="Standard Error (Cumulative)"
label(data$neuroqol_nqftg07_qposition)="Question Position"
label(data$neuroqol_nqftg10)="In the past 7 days I was frustrated by being too tired to do the things I wanted to do"
label(data$neuroqol_nqftg10_tscore)="T-score (Cumulative)"
label(data$neuroqol_nqftg10_stderror)="Standard Error (Cumulative)"
label(data$neuroqol_nqftg10_qposition)="Question Position"
label(data$neuroqol_nqftg14)="In the past 7 days I felt tired"
label(data$neuroqol_nqftg14_tscore)="T-score (Cumulative)"
label(data$neuroqol_nqftg14_stderror)="Standard Error (Cumulative)"
label(data$neuroqol_nqftg14_qposition)="Question Position"
label(data$neuroqol_nqftg02)="In the past 7 days I had to limit my social activity because I was tired"
label(data$neuroqol_nqftg02_tscore)="T-score (Cumulative)"
label(data$neuroqol_nqftg02_stderror)="Standard Error (Cumulative)"
label(data$neuroqol_nqftg02_qposition)="Question Position"
label(data$neuroqol_nqftg01)="In the past 7 days I needed help doing my usual activities because of my fatigue"
label(data$neuroqol_nqftg01_tscore)="T-score (Cumulative)"
label(data$neuroqol_nqftg01_stderror)="Standard Error (Cumulative)"
label(data$neuroqol_nqftg01_qposition)="Question Position"
label(data$neuroqol_nqftg03)="In the past 7 days I needed to sleep during the day"
label(data$neuroqol_nqftg03_tscore)="T-score (Cumulative)"
label(data$neuroqol_nqftg03_stderror)="Standard Error (Cumulative)"
label(data$neuroqol_nqftg03_qposition)="Question Position"
label(data$neuroqol_nqftg04)="In the past 7 days I had trouble starting things because I was too tired"
label(data$neuroqol_nqftg04_tscore)="T-score (Cumulative)"
label(data$neuroqol_nqftg04_stderror)="Standard Error (Cumulative)"
label(data$neuroqol_nqftg04_qposition)="Question Position"
label(data$neuroqol_nqftg05)="In the past 7 days I had trouble finishing things because I was too tired"
label(data$neuroqol_nqftg05_tscore)="T-score (Cumulative)"
label(data$neuroqol_nqftg05_stderror)="Standard Error (Cumulative)"
label(data$neuroqol_nqftg05_qposition)="Question Position"
label(data$neuroqol_nqftg08)="In the past 7 days I was too tired to take a short walk"
label(data$neuroqol_nqftg08_tscore)="T-score (Cumulative)"
label(data$neuroqol_nqftg08_stderror)="Standard Error (Cumulative)"
label(data$neuroqol_nqftg08_qposition)="Question Position"
label(data$neuroqol_nqftg09)="In the past 7 days I was too tired to eat"
label(data$neuroqol_nqftg09_tscore)="T-score (Cumulative)"
label(data$neuroqol_nqftg09_stderror)="Standard Error (Cumulative)"
label(data$neuroqol_nqftg09_qposition)="Question Position"
label(data$neuroqol_nqftg12)="In the past 7 days I was so tired that I needed to rest during the day"
label(data$neuroqol_nqftg12_tscore)="T-score (Cumulative)"
label(data$neuroqol_nqftg12_stderror)="Standard Error (Cumulative)"
label(data$neuroqol_nqftg12_qposition)="Question Position"
label(data$neuroqol_nqftg16)="In the past 7 days I felt weak all over"
label(data$neuroqol_nqftg16_tscore)="T-score (Cumulative)"
label(data$neuroqol_nqftg16_stderror)="Standard Error (Cumulative)"
label(data$neuroqol_nqftg16_qposition)="Question Position"
label(data$neuroqol_nqftg17)="In the past 7 days I needed help doing my usual activities because of weakness"
label(data$neuroqol_nqftg17_tscore)="T-score (Cumulative)"
label(data$neuroqol_nqftg17_stderror)="Standard Error (Cumulative)"
label(data$neuroqol_nqftg17_qposition)="Question Position"
label(data$neuroqol_nqftg18)="In the past 7 days I had to limit my social activity because I was physically weak"
label(data$neuroqol_nqftg18_tscore)="T-score (Cumulative)"
label(data$neuroqol_nqftg18_stderror)="Standard Error (Cumulative)"
label(data$neuroqol_nqftg18_qposition)="Question Position"
label(data$neuroqol_nqftg20)="In the past 7 days I had to force myself to get up and do things because I was physically too weak"
label(data$neuroqol_nqftg20_tscore)="T-score (Cumulative)"
label(data$neuroqol_nqftg20_stderror)="Standard Error (Cumulative)"
label(data$neuroqol_nqftg20_qposition)="Question Position"
label(data$neuroqol_bank_v10_fatigue_complete)="Complete?"
label(data$alcohol_dependence_scale_timestamp)="Survey Timestamp"
label(data$ads_1)="1. How much did you drink the last time you drank?"
label(data$ads_2)="2. Do you often have hangovers on Sunday or Monday mornings?"
label(data$ads_3)="3. Have you had the shakes when sobering up (hands tremble, shake inside)?"
label(data$ads_4)="4. Do you get physically sick (e.g., vomit, stomach cramps) as a result of drinking?"
label(data$ads_5)="5. Have you had the DTs (delirium tremens) - that is, seen, felt or heard things not really there; felt very anxious, restless, and over excited?"
label(data$ads_6)="6. When you drink, do you stumble about, stagger, and weave?"
label(data$ads_7)="7. As a result of drinking, have you felt overly hot and sweaty (feverish)?"
label(data$ads_8)="8. As a result of drinking, have you seen things that were not really there?"
label(data$ads_9)="9. Do you panic because you fear you may not have a drink when you need it?"
label(data$ads_10)="10. Have you had blackouts (loss of memory without passing out) as a result of drinking?"
label(data$ads_11)="11. Do you carry a bottle with you or keep one close at hand?"
label(data$ads_12)="12. After a period of abstinence (not drinking), do you end up drinking heavily again?"
label(data$ads_13)="13. In the past 12 months, have you passed out as a result of drinking?"
label(data$ads_14)="14. Have you had a convulsion (fit) following a period of drinking?"
label(data$ads_15)="15. Do you drink throughout the day?"
label(data$ads_16)="16. After drinking heavily, has your thinking been fuzzy or unclear?"
label(data$ads_17)="17. As a result of drinking, have you felt your heart beating rapidly?"
label(data$ads_18)="18. Do you almost constantly think about drinking and alcohol?"
label(data$ads_19)="19. As a result of drinking, have you heard things that were not really there?"
label(data$ads_20)="20. Have you had weird and frightening sensations when drinking?"
label(data$ads_21)="21. As a result of drinking have you felt things crawling on you that were not really there (e.g., bugs, spiders)?"
label(data$ads_22)="22. With respect to blackouts (loss; of memory):"
label(data$ads_23)="23. Have you tried to cut down on your drinking failed?"
label(data$ads_24)="24. Do you gulp drinks (drink quickly?)"
label(data$ads_25)="25. After taking one or two drinks, can you usually stop?"
label(data$alcohol_dependence_scale_complete)="Complete?"
label(data$rand_36_item_sf_health_survey_instrument_version_1_timestamp)="Survey Timestamp"
label(data$rand36_1)="1. In general, would you say your health is:"
label(data$rand36_2)="2. Compared to one year ago, how would your rate your health in general now?"
label(data$rand36_3)="3. Vigorous activities, such as running, lifting heavy objects, participating in strenuous sports"
label(data$rand36_4)="4. Moderate activities, such as moving a table, pushing a vacuum cleaner, bowling, or playing golf"
label(data$rand36_5)="5. Lifting or carrying groceries"
label(data$rand36_6)="6. Climbing several flights of stairs"
label(data$rand36_7)="7. Climbing one flight of stairs"
label(data$rand36_8)="8. Bending, kneeling, or stooping"
label(data$rand36_9)="9. Walking more than a mile"
label(data$rand36_10)="10. Walking several blocks"
label(data$rand36_11)="11. Walking one block"
label(data$rand36_12)="12. Bathing or dressing yourself"
label(data$rand36_13)="13. Cut down the amount of time you spent on work or other activities"
label(data$rand36_14)="14. Accomplished less than you would like"
label(data$rand36_15)="15. Were limited in the kind of work or other activities"
label(data$rand36_16)="16. Had difficulty performing the work or other activities (for example, it took extra effort)"
label(data$rand36_17)="17. Cut down the amount of time you spent on work or other activities"
label(data$rand36_18)="18. Accomplished less than you would like"
label(data$rand36_19)="19. Didnt do work or other activities as carefully as usual"
label(data$rand36_20)="20. During the past 4 weeks, to what extent has your physical health or emotional problems interfered with your normal social activities with family, friends, neighbors, or groups? (Choose one answer.)"
label(data$rand36_21)="21. How much bodily pain have you had during the past 4 weeks? (Choose one answer.)"
label(data$rand36_22)="22. During the past 4 weeks, how much did pain interfere with your normal work (including both work outside the home and housework)? (Choose one answer.)"
label(data$rand36_23)="23. Did you feel full of pep?"
label(data$rand36_24)="24. Have you been a very nervous person?"
label(data$rand36_25)="25. Have you felt so down in the dumps that nothing could cheer you up?"
label(data$rand36_26)="26. Have you felt calm and peaceful?"
label(data$rand36_27)="27. Did you have a lot of energy?"
label(data$rand36_28)="28. Have you felt downhearted and blue?"
label(data$rand36_29)="29. Did you feel worn out?"
label(data$rand36_30)="30. Have you been a happy person?"
label(data$rand36_31)="31. Did you feel tired?"
label(data$rand36_32)="32. During the past 4 weeks, how much of the time has your physical health or emotional problems interfered with your social activities (like visiting with friends, relatives, etc.)?(Choose one number)"
label(data$rand36_33)="33. I seem to get sick a little easier than other people "
label(data$rand36_34)="34. I am as healthy as anybody I know"
label(data$rand36_35)="35. I expect my health to get worse"
label(data$rand36_36)="36. My health is excellent"
label(data$rand_36_item_sf_health_survey_instrument_version_1_complete)="Complete?"
#Setting Units


#Setting Factors(will create new variable for factors)
data$residence.factor = factor(data$residence,levels=c("1","0"))
data$sex_diff_bool.factor = factor(data$sex_diff_bool,levels=c("1","0"))
data$intro_complete.factor = factor(data$intro_complete,levels=c("0","1","2"))
data$consentb.factor = factor(data$consentb,levels=c("1","2"))
data$consentb_2.factor = factor(data$consentb_2,levels=c("1","2"))
data$consentb_3.factor = factor(data$consentb_3,levels=c("1","2"))
data$consent_form_complete.factor = factor(data$consent_form_complete,levels=c("0","1","2"))
data$state.factor = factor(data$state,levels=c("1","2","3","4","5","6","7","8"))
data$year_of_birth.factor = factor(data$year_of_birth,levels=c("1921","1922","1923","1924","1925","1926","1927","1928","1929","1930","1931","1932","1933","1934","1935","1936","1937","1938","1939","1940","1941","1942","1943","1944","1945","1946","1947","1948","1949","1950","1951","1952","1953","1954","1955","1956","1957","1958","1959","1960","1961","1962","1963","1964","1965","1966","1967","1968","1969","1970","1971","1972","1973","1974","1975","1976","1977","1978","1979","1980","1981","1982","1983","1984","1985","1986","1987","1988","1989","1990","1991","1992","1993","1994","1995","1996","1997","1998","1999","2000","2001","2002"))
data$rel_status.factor = factor(data$rel_status,levels=c("1","2","3","4","5","6","7","8"))
data$home.factor = factor(data$home,levels=c("1","2","3","4","5"))
data$education.factor = factor(data$education,levels=c("1","2","3","4","5","6","7","8","9"))
data$employment_status.factor = factor(data$employment_status,levels=c("1","2","3","4","5","6"))
data$gender.factor = factor(data$gender,levels=c("1","2","3","4","5","6"))
data$ethnicity.factor = factor(data$ethnicity,levels=c("1","2","3","4","5","6","7","8","9","10","11"))
data$lgbtq.factor = factor(data$lgbtq,levels=c("1","0"))
data$mental_id.factor = factor(data$mental_id,levels=c("1","0"))
data$alcohol.factor = factor(data$alcohol,levels=c("1","0"))
data$smoke.factor = factor(data$smoke,levels=c("1","0"))
data$diseasestat.factor = factor(data$diseasestat,levels=c("1","0"))
data$demographics_complete.factor = factor(data$demographics_complete,levels=c("0","1","2"))
data$self.factor = factor(data$self,levels=c("1","2"))
data$formal.factor = factor(data$formal,levels=c("1","2"))
data$autoimmune_id___1.factor = factor(data$autoimmune_id___1,levels=c("0","1"))
data$autoimmune_id___2.factor = factor(data$autoimmune_id___2,levels=c("0","1"))
data$autoimmune_id___3.factor = factor(data$autoimmune_id___3,levels=c("0","1"))
data$autoimmune_id___4.factor = factor(data$autoimmune_id___4,levels=c("0","1"))
data$autoimmune_id___5.factor = factor(data$autoimmune_id___5,levels=c("0","1"))
data$autoimmune_id___6.factor = factor(data$autoimmune_id___6,levels=c("0","1"))
data$autoimmune_id___7.factor = factor(data$autoimmune_id___7,levels=c("0","1"))
data$autoimmune_id___8.factor = factor(data$autoimmune_id___8,levels=c("0","1"))
data$autoimmune_id___9.factor = factor(data$autoimmune_id___9,levels=c("0","1"))
data$autoimmune_id___10.factor = factor(data$autoimmune_id___10,levels=c("0","1"))
data$autoimmune_id___11.factor = factor(data$autoimmune_id___11,levels=c("0","1"))
data$autoimmune_id___12.factor = factor(data$autoimmune_id___12,levels=c("0","1"))
data$autoimmune_id___13.factor = factor(data$autoimmune_id___13,levels=c("0","1"))
data$autoimmune_id___14.factor = factor(data$autoimmune_id___14,levels=c("0","1"))
data$autoimmune_id___15.factor = factor(data$autoimmune_id___15,levels=c("0","1"))
data$autoimmune_id___16.factor = factor(data$autoimmune_id___16,levels=c("0","1"))
data$autoimmune_id___17.factor = factor(data$autoimmune_id___17,levels=c("0","1"))
data$autoimmune_id___18.factor = factor(data$autoimmune_id___18,levels=c("0","1"))
data$autoimmune_id___19.factor = factor(data$autoimmune_id___19,levels=c("0","1"))
data$autoimmune_id___20.factor = factor(data$autoimmune_id___20,levels=c("0","1"))
data$autoimmune_id___21.factor = factor(data$autoimmune_id___21,levels=c("0","1"))
data$autoimmune_id___22.factor = factor(data$autoimmune_id___22,levels=c("0","1"))
data$autoimmune_id___23.factor = factor(data$autoimmune_id___23,levels=c("0","1"))
data$autoimmune_id___24.factor = factor(data$autoimmune_id___24,levels=c("0","1"))
data$autoimmune_id___25.factor = factor(data$autoimmune_id___25,levels=c("0","1"))
data$autoimmune_id___26.factor = factor(data$autoimmune_id___26,levels=c("0","1"))
data$autoimmune_id___27.factor = factor(data$autoimmune_id___27,levels=c("0","1"))
data$autoimmune_id___28.factor = factor(data$autoimmune_id___28,levels=c("0","1"))
data$autoimmune_id___29.factor = factor(data$autoimmune_id___29,levels=c("0","1"))
data$autoimmune_id___30.factor = factor(data$autoimmune_id___30,levels=c("0","1"))
data$autoimmune_id___31.factor = factor(data$autoimmune_id___31,levels=c("0","1"))
data$autoimmune_id___32.factor = factor(data$autoimmune_id___32,levels=c("0","1"))
data$autoimmune_id___33.factor = factor(data$autoimmune_id___33,levels=c("0","1"))
data$autoimmune_id___34.factor = factor(data$autoimmune_id___34,levels=c("0","1"))
data$autoimmune_id___35.factor = factor(data$autoimmune_id___35,levels=c("0","1"))
data$autoimmune_id___36.factor = factor(data$autoimmune_id___36,levels=c("0","1"))
data$autoimmune_id___37.factor = factor(data$autoimmune_id___37,levels=c("0","1"))
data$autoimmune_id___38.factor = factor(data$autoimmune_id___38,levels=c("0","1"))
data$autoimmune_id___39.factor = factor(data$autoimmune_id___39,levels=c("0","1"))
data$autoimmune_id___40.factor = factor(data$autoimmune_id___40,levels=c("0","1"))
data$autoimmune_id___41.factor = factor(data$autoimmune_id___41,levels=c("0","1"))
data$autoimmune_id___42.factor = factor(data$autoimmune_id___42,levels=c("0","1"))
data$autoimmune_id___43.factor = factor(data$autoimmune_id___43,levels=c("0","1"))
data$autoimmune_id___44.factor = factor(data$autoimmune_id___44,levels=c("0","1"))
data$autoimmune_id___45.factor = factor(data$autoimmune_id___45,levels=c("0","1"))
data$diagnosis_type.factor = factor(data$diagnosis_type,levels=c("1","2"))
data$system.factor = factor(data$system,levels=c("1","2"))
data$diagnosis_length.factor = factor(data$diagnosis_length,levels=c("2","3","4","5","6","7","8","9","10"))
data$viral1.factor = factor(data$viral1,levels=c("1","2","3"))
data$preg1.factor = factor(data$preg1,levels=c("1","2","3"))
data$preg2.factor = factor(data$preg2,levels=c("1","2","3"))
data$meno1.factor = factor(data$meno1,levels=c("1","2","3"))
data$ill1.factor = factor(data$ill1,levels=c("1","2","3"))
data$otherlife.factor = factor(data$otherlife,levels=c("1","2","3"))
data$misdiag.factor = factor(data$misdiag,levels=c("1","0"))
data$chronic___1.factor = factor(data$chronic___1,levels=c("0","1"))
data$chronic___2.factor = factor(data$chronic___2,levels=c("0","1"))
data$chronic___3.factor = factor(data$chronic___3,levels=c("0","1"))
data$chronic___4.factor = factor(data$chronic___4,levels=c("0","1"))
data$chronic___5.factor = factor(data$chronic___5,levels=c("0","1"))
data$chronic___7.factor = factor(data$chronic___7,levels=c("0","1"))
data$chronic___8.factor = factor(data$chronic___8,levels=c("0","1"))
data$disease_and_diagnosis_complete.factor = factor(data$disease_and_diagnosis_complete,levels=c("0","1","2"))
data$fatigue.factor = factor(data$fatigue,levels=c("0","1","2","3","4","5"))
data$joint.factor = factor(data$joint,levels=c("0","1","2","3","4","5"))
data$skin.factor = factor(data$skin,levels=c("0","1","2","3","4","5"))
data$stomach.factor = factor(data$stomach,levels=c("0","1","2","3","4","5"))
data$fever.factor = factor(data$fever,levels=c("0","1","2","3","4","5"))
data$glands.factor = factor(data$glands,levels=c("0","1","2","3","4","5"))
data$chills.factor = factor(data$chills,levels=c("0","1","2","3","4","5"))
data$memory.factor = factor(data$memory,levels=c("0","1","2","3","4","5"))
data$conc.factor = factor(data$conc,levels=c("0","1","2","3","4","5"))
data$hair.factor = factor(data$hair,levels=c("0","1","2","3","4","5"))
data$skin2.factor = factor(data$skin2,levels=c("0","1","2","3","4","5"))
data$weak.factor = factor(data$weak,levels=c("0","1","2","3","4","5"))
data$pain.factor = factor(data$pain,levels=c("0","1","2","3","4","5"))
data$bruise.factor = factor(data$bruise,levels=c("0","1","2","3","4","5"))
data$treatments___1.factor = factor(data$treatments___1,levels=c("0","1"))
data$treatments___2.factor = factor(data$treatments___2,levels=c("0","1"))
data$treatments___3.factor = factor(data$treatments___3,levels=c("0","1"))
data$treatments___5.factor = factor(data$treatments___5,levels=c("0","1"))
data$treatments___4.factor = factor(data$treatments___4,levels=c("0","1"))
data$flares___1.factor = factor(data$flares___1,levels=c("0","1"))
data$flares___2.factor = factor(data$flares___2,levels=c("0","1"))
data$flares___3.factor = factor(data$flares___3,levels=c("0","1"))
data$flares___4.factor = factor(data$flares___4,levels=c("0","1"))
data$flares___6.factor = factor(data$flares___6,levels=c("0","1"))
data$flares___5.factor = factor(data$flares___5,levels=c("0","1"))
data$drug___1.factor = factor(data$drug___1,levels=c("0","1"))
data$drug___2.factor = factor(data$drug___2,levels=c("0","1"))
data$drug___3.factor = factor(data$drug___3,levels=c("0","1"))
data$drug___4.factor = factor(data$drug___4,levels=c("0","1"))
data$drug___5.factor = factor(data$drug___5,levels=c("0","1"))
data$med_other.factor = factor(data$med_other,levels=c("1","0"))
data$drug_other.factor = factor(data$drug_other,levels=c("1","0"))
data$hormone___1.factor = factor(data$hormone___1,levels=c("0","1"))
data$hormone___3.factor = factor(data$hormone___3,levels=c("0","1"))
data$hormone___2.factor = factor(data$hormone___2,levels=c("0","1"))
data$hormone___4.factor = factor(data$hormone___4,levels=c("0","1"))
data$hormone___5.factor = factor(data$hormone___5,levels=c("0","1"))
data$hormone___6.factor = factor(data$hormone___6,levels=c("0","1"))
data$hormone___7.factor = factor(data$hormone___7,levels=c("0","1"))
data$diet___1.factor = factor(data$diet___1,levels=c("0","1"))
data$diet___2.factor = factor(data$diet___2,levels=c("0","1"))
data$diet___3.factor = factor(data$diet___3,levels=c("0","1"))
data$diet___4.factor = factor(data$diet___4,levels=c("0","1"))
data$diet___5.factor = factor(data$diet___5,levels=c("0","1"))
data$diet___6.factor = factor(data$diet___6,levels=c("0","1"))
data$diet___7.factor = factor(data$diet___7,levels=c("0","1"))
data$diet___8.factor = factor(data$diet___8,levels=c("0","1"))
data$physical___1.factor = factor(data$physical___1,levels=c("0","1"))
data$physical___2.factor = factor(data$physical___2,levels=c("0","1"))
data$physical___3.factor = factor(data$physical___3,levels=c("0","1"))
data$physical___4.factor = factor(data$physical___4,levels=c("0","1"))
data$physical___5.factor = factor(data$physical___5,levels=c("0","1"))
data$specialist___1.factor = factor(data$specialist___1,levels=c("0","1"))
data$specialist___2.factor = factor(data$specialist___2,levels=c("0","1"))
data$specialist___3.factor = factor(data$specialist___3,levels=c("0","1"))
data$specialist___4.factor = factor(data$specialist___4,levels=c("0","1"))
data$specialist___5.factor = factor(data$specialist___5,levels=c("0","1"))
data$specialist___6.factor = factor(data$specialist___6,levels=c("0","1"))
data$specialist___7.factor = factor(data$specialist___7,levels=c("0","1"))
data$specialist___8.factor = factor(data$specialist___8,levels=c("0","1"))
data$specialist___9.factor = factor(data$specialist___9,levels=c("0","1"))
data$specialist___10.factor = factor(data$specialist___10,levels=c("0","1"))
data$therapy___1.factor = factor(data$therapy___1,levels=c("0","1"))
data$therapy___2.factor = factor(data$therapy___2,levels=c("0","1"))
data$therapy___3.factor = factor(data$therapy___3,levels=c("0","1"))
data$therapy___4.factor = factor(data$therapy___4,levels=c("0","1"))
data$therapy___5.factor = factor(data$therapy___5,levels=c("0","1"))
data$therapy___6.factor = factor(data$therapy___6,levels=c("0","1"))
data$therapy___7.factor = factor(data$therapy___7,levels=c("0","1"))
data$therapy___8.factor = factor(data$therapy___8,levels=c("0","1"))
data$symptom_management_complete.factor = factor(data$symptom_management_complete,levels=c("0","1","2"))
data$selfmental.factor = factor(data$selfmental,levels=c("1","2"))
data$formalmental.factor = factor(data$formalmental,levels=c("1","2"))
data$mental_diagn___1.factor = factor(data$mental_diagn___1,levels=c("0","1"))
data$mental_diagn___2.factor = factor(data$mental_diagn___2,levels=c("0","1"))
data$mental_diagn___3.factor = factor(data$mental_diagn___3,levels=c("0","1"))
data$mental_diagn___4.factor = factor(data$mental_diagn___4,levels=c("0","1"))
data$mental_diagn___5.factor = factor(data$mental_diagn___5,levels=c("0","1"))
data$mental_diagn___6.factor = factor(data$mental_diagn___6,levels=c("0","1"))
data$mental_diagn___7.factor = factor(data$mental_diagn___7,levels=c("0","1"))
data$ptsd___1.factor = factor(data$ptsd___1,levels=c("0","1"))
data$ptsd___2.factor = factor(data$ptsd___2,levels=c("0","1"))
data$ptsd___3.factor = factor(data$ptsd___3,levels=c("0","1"))
data$ptsd___4.factor = factor(data$ptsd___4,levels=c("0","1"))
data$ptsd___5.factor = factor(data$ptsd___5,levels=c("0","1"))
data$ptsd___7.factor = factor(data$ptsd___7,levels=c("0","1"))
data$ptsd___6.factor = factor(data$ptsd___6,levels=c("0","1"))
data$ptsd_freq.factor = factor(data$ptsd_freq,levels=c("1","2","3","4","5","6"))
data$selffamment.factor = factor(data$selffamment,levels=c("1","2"))
data$formalfamment.factor = factor(data$formalfamment,levels=c("1","2"))
data$mental_diagn_fam___1.factor = factor(data$mental_diagn_fam___1,levels=c("0","1"))
data$mental_diagn_fam___2.factor = factor(data$mental_diagn_fam___2,levels=c("0","1"))
data$mental_diagn_fam___3.factor = factor(data$mental_diagn_fam___3,levels=c("0","1"))
data$mental_diagn_fam___4.factor = factor(data$mental_diagn_fam___4,levels=c("0","1"))
data$mental_diagn_fam___5.factor = factor(data$mental_diagn_fam___5,levels=c("0","1"))
data$mental_diagn_fam___6.factor = factor(data$mental_diagn_fam___6,levels=c("0","1"))
data$mental_diagn_fam___7.factor = factor(data$mental_diagn_fam___7,levels=c("0","1"))
data$ptsd_fam___1.factor = factor(data$ptsd_fam___1,levels=c("0","1"))
data$ptsd_fam___2.factor = factor(data$ptsd_fam___2,levels=c("0","1"))
data$ptsd_fam___3.factor = factor(data$ptsd_fam___3,levels=c("0","1"))
data$ptsd_fam___4.factor = factor(data$ptsd_fam___4,levels=c("0","1"))
data$ptsd_fam___5.factor = factor(data$ptsd_fam___5,levels=c("0","1"))
data$ptsd_fam___6.factor = factor(data$ptsd_fam___6,levels=c("0","1"))
data$ptsd_fam___8.factor = factor(data$ptsd_fam___8,levels=c("0","1"))
data$ptsd_fam___7.factor = factor(data$ptsd_fam___7,levels=c("0","1"))
data$mental_health_complete.factor = factor(data$mental_health_complete,levels=c("0","1","2"))
data$neuroqol_nqftg13.factor = factor(data$neuroqol_nqftg13,levels=c("1","2","3","4","5"))
data$neuroqol_nqftg11.factor = factor(data$neuroqol_nqftg11,levels=c("1","2","3","4","5"))
data$neuroqol_nqftg15.factor = factor(data$neuroqol_nqftg15,levels=c("1","2","3","4","5"))
data$neuroqol_nqftg06.factor = factor(data$neuroqol_nqftg06,levels=c("1","2","3","4","5"))
data$neuroqol_nqftg07.factor = factor(data$neuroqol_nqftg07,levels=c("1","2","3","4","5"))
data$neuroqol_nqftg10.factor = factor(data$neuroqol_nqftg10,levels=c("1","2","3","4","5"))
data$neuroqol_nqftg14.factor = factor(data$neuroqol_nqftg14,levels=c("1","2","3","4","5"))
data$neuroqol_nqftg02.factor = factor(data$neuroqol_nqftg02,levels=c("1","2","3","4","5"))
data$neuroqol_nqftg01.factor = factor(data$neuroqol_nqftg01,levels=c("1","2","3","4","5"))
data$neuroqol_nqftg03.factor = factor(data$neuroqol_nqftg03,levels=c("1","2","3","4","5"))
data$neuroqol_nqftg04.factor = factor(data$neuroqol_nqftg04,levels=c("1","2","3","4","5"))
data$neuroqol_nqftg05.factor = factor(data$neuroqol_nqftg05,levels=c("1","2","3","4","5"))
data$neuroqol_nqftg08.factor = factor(data$neuroqol_nqftg08,levels=c("1","2","3","4","5"))
data$neuroqol_nqftg09.factor = factor(data$neuroqol_nqftg09,levels=c("1","2","3","4","5"))
data$neuroqol_nqftg12.factor = factor(data$neuroqol_nqftg12,levels=c("1","2","3","4","5"))
data$neuroqol_nqftg16.factor = factor(data$neuroqol_nqftg16,levels=c("1","2","3","4","5"))
data$neuroqol_nqftg17.factor = factor(data$neuroqol_nqftg17,levels=c("1","2","3","4","5"))
data$neuroqol_nqftg18.factor = factor(data$neuroqol_nqftg18,levels=c("1","2","3","4","5"))
data$neuroqol_nqftg20.factor = factor(data$neuroqol_nqftg20,levels=c("1","2","3","4","5"))
data$neuroqol_bank_v10_fatigue_complete.factor = factor(data$neuroqol_bank_v10_fatigue_complete,levels=c("0","1","2"))
data$ads_1.factor = factor(data$ads_1,levels=c("0","1","2"))
data$ads_2.factor = factor(data$ads_2,levels=c("0","1"))
data$ads_3.factor = factor(data$ads_3,levels=c("0","1","2"))
data$ads_4.factor = factor(data$ads_4,levels=c("0","1","2"))
data$ads_5.factor = factor(data$ads_5,levels=c("0","1","2"))
data$ads_6.factor = factor(data$ads_6,levels=c("0","1","2"))
data$ads_7.factor = factor(data$ads_7,levels=c("0","1","2"))
data$ads_8.factor = factor(data$ads_8,levels=c("0","1","2"))
data$ads_9.factor = factor(data$ads_9,levels=c("0","1"))
data$ads_10.factor = factor(data$ads_10,levels=c("0","1","2","3"))
data$ads_11.factor = factor(data$ads_11,levels=c("0","1","2"))
data$ads_12.factor = factor(data$ads_12,levels=c("0","1","2"))
data$ads_13.factor = factor(data$ads_13,levels=c("0","1","2"))
data$ads_14.factor = factor(data$ads_14,levels=c("0","1","2"))
data$ads_15.factor = factor(data$ads_15,levels=c("0","1"))
data$ads_16.factor = factor(data$ads_16,levels=c("0","1","2","3"))
data$ads_17.factor = factor(data$ads_17,levels=c("0","1","2"))
data$ads_18.factor = factor(data$ads_18,levels=c("0","1"))
data$ads_19.factor = factor(data$ads_19,levels=c("0","1","2"))
data$ads_20.factor = factor(data$ads_20,levels=c("0","1","2"))
data$ads_21.factor = factor(data$ads_21,levels=c("0","1","2"))
data$ads_22.factor = factor(data$ads_22,levels=c("0","1","2","3"))
data$ads_23.factor = factor(data$ads_23,levels=c("0","1","2"))
data$ads_24.factor = factor(data$ads_24,levels=c("0","1"))
data$ads_25.factor = factor(data$ads_25,levels=c("0","1"))
data$alcohol_dependence_scale_complete.factor = factor(data$alcohol_dependence_scale_complete,levels=c("0","1","2"))
data$rand36_1.factor = factor(data$rand36_1,levels=c("100","75","50","25","0"))
data$rand36_2.factor = factor(data$rand36_2,levels=c("100","75","50","25","0"))
data$rand36_3.factor = factor(data$rand36_3,levels=c("0","50","100"))
data$rand36_4.factor = factor(data$rand36_4,levels=c("0","50","100"))
data$rand36_5.factor = factor(data$rand36_5,levels=c("0","50","100"))
data$rand36_6.factor = factor(data$rand36_6,levels=c("0","50","100"))
data$rand36_7.factor = factor(data$rand36_7,levels=c("0","50","100"))
data$rand36_8.factor = factor(data$rand36_8,levels=c("0","50","100"))
data$rand36_9.factor = factor(data$rand36_9,levels=c("0","50","100"))
data$rand36_10.factor = factor(data$rand36_10,levels=c("0","50","100"))
data$rand36_11.factor = factor(data$rand36_11,levels=c("0","50","100"))
data$rand36_12.factor = factor(data$rand36_12,levels=c("0","50","100"))
data$rand36_13.factor = factor(data$rand36_13,levels=c("0","100"))
data$rand36_14.factor = factor(data$rand36_14,levels=c("0","100"))
data$rand36_15.factor = factor(data$rand36_15,levels=c("0","100"))
data$rand36_16.factor = factor(data$rand36_16,levels=c("0","100"))
data$rand36_17.factor = factor(data$rand36_17,levels=c("0","100"))
data$rand36_18.factor = factor(data$rand36_18,levels=c("0","100"))
data$rand36_19.factor = factor(data$rand36_19,levels=c("0","100"))
data$rand36_20.factor = factor(data$rand36_20,levels=c("100","75","50","25","0"))
data$rand36_21.factor = factor(data$rand36_21,levels=c("100","80","60","40","20","0"))
data$rand36_22.factor = factor(data$rand36_22,levels=c("100","75","50","25","0"))
data$rand36_23.factor = factor(data$rand36_23,levels=c("100","80","60","40","20","0"))
data$rand36_24.factor = factor(data$rand36_24,levels=c("0","20","40","60","80","100"))
data$rand36_25.factor = factor(data$rand36_25,levels=c("0","20","40","60","80","100"))
data$rand36_26.factor = factor(data$rand36_26,levels=c("100","80","60","40","20","0"))
data$rand36_27.factor = factor(data$rand36_27,levels=c("100","80","60","40","20","0"))
data$rand36_28.factor = factor(data$rand36_28,levels=c("0","20","40","60","80","100"))
data$rand36_29.factor = factor(data$rand36_29,levels=c("0","20","40","60","80","100"))
data$rand36_30.factor = factor(data$rand36_30,levels=c("100","80","60","40","20","0"))
data$rand36_31.factor = factor(data$rand36_31,levels=c("0","20","40","60","80","100"))
data$rand36_32.factor = factor(data$rand36_32,levels=c("0","25","50","75","100"))
data$rand36_33.factor = factor(data$rand36_33,levels=c("0","25","50","75","100"))
data$rand36_34.factor = factor(data$rand36_34,levels=c("100","75","50","25","0"))
data$rand36_35.factor = factor(data$rand36_35,levels=c("0","25","50","75","100"))
data$rand36_36.factor = factor(data$rand36_36,levels=c("100","75","50","25","0"))
data$rand_36_item_sf_health_survey_instrument_version_1_complete.factor = factor(data$rand_36_item_sf_health_survey_instrument_version_1_complete,levels=c("0","1","2"))

levels(data$residence.factor)=c("Yes","No")
levels(data$sex_diff_bool.factor)=c("Yes","No")
levels(data$intro_complete.factor)=c("Incomplete","Unverified","Complete")
levels(data$consentb.factor)=c("this specific research project only","this research project and other research that is closely related to this research project")
levels(data$consentb_2.factor)=c("Yes","No")
levels(data$consentb_3.factor)=c("Yes","No")
levels(data$consent_form_complete.factor)=c("Incomplete","Unverified","Complete")
levels(data$state.factor)=c("ACT","NSW","NT","QLD","SA","TAS","VIC","WA")
levels(data$year_of_birth.factor)=c("1921","1922","1923","1924","1925","1926","1927","1928","1929","1930","1931","1932","1933","1934","1935","1936","1937","1938","1939","1940","1941","1942","1943","1944","1945","1946","1947","1948","1949","1950","1951","1952","1953","1954","1955","1956","1957","1958","1959","1960","1961","1962","1963","1964","1965","1966","1967","1968","1969","1970","1971","1972","1973","1974","1975","1976","1977","1978","1979","1980","1981","1982","1983","1984","1985","1986","1987","1988","1989","1990","1991","1992","1993","1994","1995","1996","1997","1998","1999","2000","2001","2002")
levels(data$rel_status.factor)=c("Married","Widowed","Divorced","Separated","In a domestic partnership or civil union","Single but cohabiting with a significant other","Single never married","Prefer not to answer")
levels(data$home.factor)=c("Owned or being bought by you or someone in your household","Rented","Occupied without payment of rent (e.g., with carer or parents)","Homeless or squatting","Other")
levels(data$education.factor)=c("Departed prior to school certificate completion (Year 10 equivalent)","School certificate (Year 10 equivalent)","High school degree or equivalent (Year 12 equivalent)","Some college but no degree","Associate degree","Bachelor degree","Masters degree","Doctorate (PhD)","Prefer not to answer")
levels(data$employment_status.factor)=c("Employed working 40 or more hours per week","Employed working 1-39 hours per week","Not employed looking for work","Not employed NOT looking for work","Retired","Disabled not able to work")
levels(data$gender.factor)=c("Female","Male","Gender fluid","Non-binary","Other","Prefer not to say")
levels(data$ethnicity.factor)=c("Aboriginal and Torres Strait Islander or Descent","African American or African Origin or Descent","American Indian or Alaska Native Origin or Descent","Asian Origin or Descent","European Origin or Descent","Hispanic or Latino/a Origin or Descent","Middle Eastern Origin or Descent","Native Hawaiian or other","Pacific Islander Origin or Descent","None of these fully describe me","Prefer not to answer")
levels(data$lgbtq.factor)=c("Yes","No")
levels(data$mental_id.factor)=c("Yes","No")
levels(data$alcohol.factor)=c("Yes","No")
levels(data$smoke.factor)=c("Yes","No")
levels(data$diseasestat.factor)=c("Yes","No")
levels(data$demographics_complete.factor)=c("Incomplete","Unverified","Complete")
levels(data$self.factor)=c("Yes","No")
levels(data$formal.factor)=c("Yes","No")
levels(data$autoimmune_id___1.factor)=c("Unchecked","Checked")
levels(data$autoimmune_id___2.factor)=c("Unchecked","Checked")
levels(data$autoimmune_id___3.factor)=c("Unchecked","Checked")
levels(data$autoimmune_id___4.factor)=c("Unchecked","Checked")
levels(data$autoimmune_id___5.factor)=c("Unchecked","Checked")
levels(data$autoimmune_id___6.factor)=c("Unchecked","Checked")
levels(data$autoimmune_id___7.factor)=c("Unchecked","Checked")
levels(data$autoimmune_id___8.factor)=c("Unchecked","Checked")
levels(data$autoimmune_id___9.factor)=c("Unchecked","Checked")
levels(data$autoimmune_id___10.factor)=c("Unchecked","Checked")
levels(data$autoimmune_id___11.factor)=c("Unchecked","Checked")
levels(data$autoimmune_id___12.factor)=c("Unchecked","Checked")
levels(data$autoimmune_id___13.factor)=c("Unchecked","Checked")
levels(data$autoimmune_id___14.factor)=c("Unchecked","Checked")
levels(data$autoimmune_id___15.factor)=c("Unchecked","Checked")
levels(data$autoimmune_id___16.factor)=c("Unchecked","Checked")
levels(data$autoimmune_id___17.factor)=c("Unchecked","Checked")
levels(data$autoimmune_id___18.factor)=c("Unchecked","Checked")
levels(data$autoimmune_id___19.factor)=c("Unchecked","Checked")
levels(data$autoimmune_id___20.factor)=c("Unchecked","Checked")
levels(data$autoimmune_id___21.factor)=c("Unchecked","Checked")
levels(data$autoimmune_id___22.factor)=c("Unchecked","Checked")
levels(data$autoimmune_id___23.factor)=c("Unchecked","Checked")
levels(data$autoimmune_id___24.factor)=c("Unchecked","Checked")
levels(data$autoimmune_id___25.factor)=c("Unchecked","Checked")
levels(data$autoimmune_id___26.factor)=c("Unchecked","Checked")
levels(data$autoimmune_id___27.factor)=c("Unchecked","Checked")
levels(data$autoimmune_id___28.factor)=c("Unchecked","Checked")
levels(data$autoimmune_id___29.factor)=c("Unchecked","Checked")
levels(data$autoimmune_id___30.factor)=c("Unchecked","Checked")
levels(data$autoimmune_id___31.factor)=c("Unchecked","Checked")
levels(data$autoimmune_id___32.factor)=c("Unchecked","Checked")
levels(data$autoimmune_id___33.factor)=c("Unchecked","Checked")
levels(data$autoimmune_id___34.factor)=c("Unchecked","Checked")
levels(data$autoimmune_id___35.factor)=c("Unchecked","Checked")
levels(data$autoimmune_id___36.factor)=c("Unchecked","Checked")
levels(data$autoimmune_id___37.factor)=c("Unchecked","Checked")
levels(data$autoimmune_id___38.factor)=c("Unchecked","Checked")
levels(data$autoimmune_id___39.factor)=c("Unchecked","Checked")
levels(data$autoimmune_id___40.factor)=c("Unchecked","Checked")
levels(data$autoimmune_id___41.factor)=c("Unchecked","Checked")
levels(data$autoimmune_id___42.factor)=c("Unchecked","Checked")
levels(data$autoimmune_id___43.factor)=c("Unchecked","Checked")
levels(data$autoimmune_id___44.factor)=c("Unchecked","Checked")
levels(data$autoimmune_id___45.factor)=c("Unchecked","Checked")
levels(data$diagnosis_type.factor)=c("GP","Specialist")
levels(data$system.factor)=c("Public medical system (Medicare benefits schedule, public hospital, etc.)","Private medical system")
levels(data$diagnosis_length.factor)=c("Less than six months","6 months - 1 year","1 - 2 years","3 - 4 years","4 - 5 years","6 - 7  years","7 - 8  years","9 - 10 years","10+  years (please specify length in the other section)")
levels(data$viral1.factor)=c("Yes","No","Unknown")
levels(data$preg1.factor)=c("Yes","No","Unknown")
levels(data$preg2.factor)=c("Yes","No","Unknown")
levels(data$meno1.factor)=c("Yes","No","Unknown")
levels(data$ill1.factor)=c("Yes","No","Unknown")
levels(data$otherlife.factor)=c("Yes","No","Unknown")
levels(data$misdiag.factor)=c("Yes","No")
levels(data$chronic___1.factor)=c("Unchecked","Checked")
levels(data$chronic___2.factor)=c("Unchecked","Checked")
levels(data$chronic___3.factor)=c("Unchecked","Checked")
levels(data$chronic___4.factor)=c("Unchecked","Checked")
levels(data$chronic___5.factor)=c("Unchecked","Checked")
levels(data$chronic___7.factor)=c("Unchecked","Checked")
levels(data$chronic___8.factor)=c("Unchecked","Checked")
levels(data$disease_and_diagnosis_complete.factor)=c("Incomplete","Unverified","Complete")
levels(data$fatigue.factor)=c("None","Very mild","Mild","Moderate","Severe","Very severe")
levels(data$joint.factor)=c("None","Very mild","Mild","Moderate","Severe","Very severe")
levels(data$skin.factor)=c("None","Very mild","Mild","Moderate","Severe","Very severe")
levels(data$stomach.factor)=c("None","Very mild","Mild","Moderate","Severe","Very severe")
levels(data$fever.factor)=c("None","Very mild","Mild","Moderate","Severe","Very severe")
levels(data$glands.factor)=c("None","Very mild","Mild","Moderate","Severe","Very severe")
levels(data$chills.factor)=c("None","Very mild","Mild","Moderate","Severe","Very severe")
levels(data$memory.factor)=c("None","Very mild","Mild","Moderate","Severe","Very severe")
levels(data$conc.factor)=c("None","Very mild","Mild","Moderate","Severe","Very severe")
levels(data$hair.factor)=c("None","Very mild","Mild","Moderate","Severe","Very severe")
levels(data$skin2.factor)=c("None","Very mild","Mild","Moderate","Severe","Very severe")
levels(data$weak.factor)=c("None","Very mild","Mild","Moderate","Severe","Very severe")
levels(data$pain.factor)=c("None","Very mild","Mild","Moderate","Severe","Very severe")
levels(data$bruise.factor)=c("None","Very mild","Mild","Moderate","Severe","Very severe")
levels(data$treatments___1.factor)=c("Unchecked","Checked")
levels(data$treatments___2.factor)=c("Unchecked","Checked")
levels(data$treatments___3.factor)=c("Unchecked","Checked")
levels(data$treatments___5.factor)=c("Unchecked","Checked")
levels(data$treatments___4.factor)=c("Unchecked","Checked")
levels(data$flares___1.factor)=c("Unchecked","Checked")
levels(data$flares___2.factor)=c("Unchecked","Checked")
levels(data$flares___3.factor)=c("Unchecked","Checked")
levels(data$flares___4.factor)=c("Unchecked","Checked")
levels(data$flares___6.factor)=c("Unchecked","Checked")
levels(data$flares___5.factor)=c("Unchecked","Checked")
levels(data$drug___1.factor)=c("Unchecked","Checked")
levels(data$drug___2.factor)=c("Unchecked","Checked")
levels(data$drug___3.factor)=c("Unchecked","Checked")
levels(data$drug___4.factor)=c("Unchecked","Checked")
levels(data$drug___5.factor)=c("Unchecked","Checked")
levels(data$med_other.factor)=c("Yes","No")
levels(data$drug_other.factor)=c("Yes","No")
levels(data$hormone___1.factor)=c("Unchecked","Checked")
levels(data$hormone___3.factor)=c("Unchecked","Checked")
levels(data$hormone___2.factor)=c("Unchecked","Checked")
levels(data$hormone___4.factor)=c("Unchecked","Checked")
levels(data$hormone___5.factor)=c("Unchecked","Checked")
levels(data$hormone___6.factor)=c("Unchecked","Checked")
levels(data$hormone___7.factor)=c("Unchecked","Checked")
levels(data$diet___1.factor)=c("Unchecked","Checked")
levels(data$diet___2.factor)=c("Unchecked","Checked")
levels(data$diet___3.factor)=c("Unchecked","Checked")
levels(data$diet___4.factor)=c("Unchecked","Checked")
levels(data$diet___5.factor)=c("Unchecked","Checked")
levels(data$diet___6.factor)=c("Unchecked","Checked")
levels(data$diet___7.factor)=c("Unchecked","Checked")
levels(data$diet___8.factor)=c("Unchecked","Checked")
levels(data$physical___1.factor)=c("Unchecked","Checked")
levels(data$physical___2.factor)=c("Unchecked","Checked")
levels(data$physical___3.factor)=c("Unchecked","Checked")
levels(data$physical___4.factor)=c("Unchecked","Checked")
levels(data$physical___5.factor)=c("Unchecked","Checked")
levels(data$specialist___1.factor)=c("Unchecked","Checked")
levels(data$specialist___2.factor)=c("Unchecked","Checked")
levels(data$specialist___3.factor)=c("Unchecked","Checked")
levels(data$specialist___4.factor)=c("Unchecked","Checked")
levels(data$specialist___5.factor)=c("Unchecked","Checked")
levels(data$specialist___6.factor)=c("Unchecked","Checked")
levels(data$specialist___7.factor)=c("Unchecked","Checked")
levels(data$specialist___8.factor)=c("Unchecked","Checked")
levels(data$specialist___9.factor)=c("Unchecked","Checked")
levels(data$specialist___10.factor)=c("Unchecked","Checked")
levels(data$therapy___1.factor)=c("Unchecked","Checked")
levels(data$therapy___2.factor)=c("Unchecked","Checked")
levels(data$therapy___3.factor)=c("Unchecked","Checked")
levels(data$therapy___4.factor)=c("Unchecked","Checked")
levels(data$therapy___5.factor)=c("Unchecked","Checked")
levels(data$therapy___6.factor)=c("Unchecked","Checked")
levels(data$therapy___7.factor)=c("Unchecked","Checked")
levels(data$therapy___8.factor)=c("Unchecked","Checked")
levels(data$symptom_management_complete.factor)=c("Incomplete","Unverified","Complete")
levels(data$selfmental.factor)=c("Yes","No")
levels(data$formalmental.factor)=c("Yes","No")
levels(data$mental_diagn___1.factor)=c("Unchecked","Checked")
levels(data$mental_diagn___2.factor)=c("Unchecked","Checked")
levels(data$mental_diagn___3.factor)=c("Unchecked","Checked")
levels(data$mental_diagn___4.factor)=c("Unchecked","Checked")
levels(data$mental_diagn___5.factor)=c("Unchecked","Checked")
levels(data$mental_diagn___6.factor)=c("Unchecked","Checked")
levels(data$mental_diagn___7.factor)=c("Unchecked","Checked")
levels(data$ptsd___1.factor)=c("Unchecked","Checked")
levels(data$ptsd___2.factor)=c("Unchecked","Checked")
levels(data$ptsd___3.factor)=c("Unchecked","Checked")
levels(data$ptsd___4.factor)=c("Unchecked","Checked")
levels(data$ptsd___5.factor)=c("Unchecked","Checked")
levels(data$ptsd___7.factor)=c("Unchecked","Checked")
levels(data$ptsd___6.factor)=c("Unchecked","Checked")
levels(data$ptsd_freq.factor)=c("Rarely","Monthly","Weekly","Daily","Hourly","Other")
levels(data$selffamment.factor)=c("Yes","No")
levels(data$formalfamment.factor)=c("Yes","No")
levels(data$mental_diagn_fam___1.factor)=c("Unchecked","Checked")
levels(data$mental_diagn_fam___2.factor)=c("Unchecked","Checked")
levels(data$mental_diagn_fam___3.factor)=c("Unchecked","Checked")
levels(data$mental_diagn_fam___4.factor)=c("Unchecked","Checked")
levels(data$mental_diagn_fam___5.factor)=c("Unchecked","Checked")
levels(data$mental_diagn_fam___6.factor)=c("Unchecked","Checked")
levels(data$mental_diagn_fam___7.factor)=c("Unchecked","Checked")
levels(data$ptsd_fam___1.factor)=c("Unchecked","Checked")
levels(data$ptsd_fam___2.factor)=c("Unchecked","Checked")
levels(data$ptsd_fam___3.factor)=c("Unchecked","Checked")
levels(data$ptsd_fam___4.factor)=c("Unchecked","Checked")
levels(data$ptsd_fam___5.factor)=c("Unchecked","Checked")
levels(data$ptsd_fam___6.factor)=c("Unchecked","Checked")
levels(data$ptsd_fam___8.factor)=c("Unchecked","Checked")
levels(data$ptsd_fam___7.factor)=c("Unchecked","Checked")
levels(data$mental_health_complete.factor)=c("Incomplete","Unverified","Complete")
levels(data$neuroqol_nqftg13.factor)=c("Never","Rarely","Sometimes","Often","Always")
levels(data$neuroqol_nqftg11.factor)=c("Never","Rarely","Sometimes","Often","Always")
levels(data$neuroqol_nqftg15.factor)=c("Never","Rarely","Sometimes","Often","Always")
levels(data$neuroqol_nqftg06.factor)=c("Never","Rarely","Sometimes","Often","Always")
levels(data$neuroqol_nqftg07.factor)=c("Never","Rarely","Sometimes","Often","Always")
levels(data$neuroqol_nqftg10.factor)=c("Never","Rarely","Sometimes","Often","Always")
levels(data$neuroqol_nqftg14.factor)=c("Never","Rarely","Sometimes","Often","Always")
levels(data$neuroqol_nqftg02.factor)=c("Never","Rarely","Sometimes","Often","Always")
levels(data$neuroqol_nqftg01.factor)=c("Never","Rarely","Sometimes","Often","Always")
levels(data$neuroqol_nqftg03.factor)=c("Never","Rarely","Sometimes","Often","Always")
levels(data$neuroqol_nqftg04.factor)=c("Never","Rarely","Sometimes","Often","Always")
levels(data$neuroqol_nqftg05.factor)=c("Never","Rarely","Sometimes","Often","Always")
levels(data$neuroqol_nqftg08.factor)=c("Never","Rarely","Sometimes","Often","Always")
levels(data$neuroqol_nqftg09.factor)=c("Never","Rarely","Sometimes","Often","Always")
levels(data$neuroqol_nqftg12.factor)=c("Never","Rarely","Sometimes","Often","Always")
levels(data$neuroqol_nqftg16.factor)=c("Never","Rarely","Sometimes","Often","Always")
levels(data$neuroqol_nqftg17.factor)=c("Never","Rarely","Sometimes","Often","Always")
levels(data$neuroqol_nqftg18.factor)=c("Never","Rarely","Sometimes","Often","Always")
levels(data$neuroqol_nqftg20.factor)=c("Never","Rarely","Sometimes","Often","Always")
levels(data$neuroqol_bank_v10_fatigue_complete.factor)=c("Incomplete","Unverified","Complete")
levels(data$ads_1.factor)=c("a. Enough to get high or less","b. Enough to get drunk","c. Enough to pass out")
levels(data$ads_2.factor)=c("a. No","b. Yes")
levels(data$ads_3.factor)=c("a. No","b. Sometimes","c. Often")
levels(data$ads_4.factor)=c("a. No","b. Sometimes","c. Almost every time I drink")
levels(data$ads_5.factor)=c("a. No","b. Sometimes","c. Several times")
levels(data$ads_6.factor)=c("a. No","b. Sometimes","c. Often")
levels(data$ads_7.factor)=c("a. No","b. Once","c. Several times")
levels(data$ads_8.factor)=c("a. No","b. Once","c. Several times")
levels(data$ads_9.factor)=c("a. No","b. Yes")
levels(data$ads_10.factor)=c("a. No, never","b. Sometimes","c. Often","d. Almost every time I drink")
levels(data$ads_11.factor)=c("a. No","b. Some of the time","c. Most of the time")
levels(data$ads_12.factor)=c("a. No","b. Sometimes","c. Almost every time I drink")
levels(data$ads_13.factor)=c("a. No","b. Once","c. More than once")
levels(data$ads_14.factor)=c("a. No","b. Yes","c. Several times")
levels(data$ads_15.factor)=c("a. No","b. Yes")
levels(data$ads_16.factor)=c("a. No","b. Yes, but only for a few hours","c. Yes, for one or two days","d. Yes, for many days")
levels(data$ads_17.factor)=c("a. No","b. Yes","c. Several times")
levels(data$ads_18.factor)=c("a. No","b. Yes")
levels(data$ads_19.factor)=c("a. No","b. Yes","c. Several times")
levels(data$ads_20.factor)=c("a. No","b. Once or twice","c. Often")
levels(data$ads_21.factor)=c("a. No","b. Yes","c. Several times")
levels(data$ads_22.factor)=c("a. Have never had a blackout","b. Have had blackouts that last less than an hour","c. Have had blackouts that last for several hours","d. Have had blackouts that last a day or more")
levels(data$ads_23.factor)=c("a. No","b. Once","c. Several times")
levels(data$ads_24.factor)=c("a. No","b. Yes")
levels(data$ads_25.factor)=c("a. Yes","b. No")
levels(data$alcohol_dependence_scale_complete.factor)=c("Incomplete","Unverified","Complete")
levels(data$rand36_1.factor)=c("1 Excellent","2 Very good","3 Good","4 Fair","5 Poor")
levels(data$rand36_2.factor)=c("1 Much better now than one year ago","2 Somewhat better now than one year ago","3 About the same","4 Somewhat worse now than one year ago","5 Much worse now than one year ago")
levels(data$rand36_3.factor)=c("1 Yes, limited a lot","2 Yes, limited a little","3 No, not limited at all")
levels(data$rand36_4.factor)=c("1 Yes, limited a lot","2 Yes, limited a little","3 No, not limited at all")
levels(data$rand36_5.factor)=c("1 Yes, limited a lot","2 Yes, limited a little","3 No, not limited at all")
levels(data$rand36_6.factor)=c("1 Yes, limited a lot","2 Yes, limited a little","3 No, not limited at all")
levels(data$rand36_7.factor)=c("1 Yes, limited a lot","2 Yes, limited a little","3 No, not limited at all")
levels(data$rand36_8.factor)=c("1 Yes, limited a lot","2 Yes, limited a little","3 No, not limited at all")
levels(data$rand36_9.factor)=c("1 Yes, limited a lot","2 Yes, limited a little","3 No, not limited at all")
levels(data$rand36_10.factor)=c("1 Yes, limited a lot","2 Yes, limited a little","3 No, not limited at all")
levels(data$rand36_11.factor)=c("1 Yes, limited a lot","2 Yes, limited a little","3 No, not limited at all")
levels(data$rand36_12.factor)=c("1 Yes, limited a lot","2 Yes, limited a little","3 No, not limited at all")
levels(data$rand36_13.factor)=c("1 Yes","2 No")
levels(data$rand36_14.factor)=c("1 Yes","2 No")
levels(data$rand36_15.factor)=c("1 Yes","2 No")
levels(data$rand36_16.factor)=c("1 Yes","2 No")
levels(data$rand36_17.factor)=c("1 Yes","2 No")
levels(data$rand36_18.factor)=c("1 Yes","2 No")
levels(data$rand36_19.factor)=c("1 Yes","2 No")
levels(data$rand36_20.factor)=c("1 Not at all","2 Slightly","3 Moderately","4 Quite a bit","5 Extremely")
levels(data$rand36_21.factor)=c("1 None","2 Very mild","3 Mild","4 Moderate","5 Severe","6 Very severe")
levels(data$rand36_22.factor)=c("1 Not at all","2 A little bit","3 Moderately","4 Quite a bit","5 Extremely")
levels(data$rand36_23.factor)=c("1 All of the time","2 Most of the time","3 A good bit of the time","4 Some of the time","5 A little of the time","6 None of the time")
levels(data$rand36_24.factor)=c("1 All of the time","2 Most of the time","3 A good bit of the time","4 Some of the time","5 A little of the time","6 None of the time")
levels(data$rand36_25.factor)=c("1 All of the time","2 Most of the time","3 A good bit of the time","4 Some of the time","5 A little of the time","6 None of the time")
levels(data$rand36_26.factor)=c("1 All of the time","2 Most of the time","3 A good bit of the time","4 Some of the time","5 A little of the time","6 None of the time")
levels(data$rand36_27.factor)=c("1 All of the time","2 Most of the time","3 A good bit of the time","4 Some of the time","5 A little of the time","6 None of the time")
levels(data$rand36_28.factor)=c("1 All of the time","2 Most of the time","3 A good bit of the time","4 Some of the time","5 A little of the time","6 None of the time")
levels(data$rand36_29.factor)=c("1 All of the time","2 Most of the time","3 A good bit of the time","4 Some of the time","5 A little of the time","6 None of the time")
levels(data$rand36_30.factor)=c("1 All of the time","2 Most of the time","3 A good bit of the time","4 Some of the time","5 A little of the time","6 None of the time")
levels(data$rand36_31.factor)=c("1 All of the time","2 Most of the time","3 A good bit of the time","4 Some of the time","5 A little of the time","6 None of the time")
levels(data$rand36_32.factor)=c("1 All of the time","2 Most of the time","3 Some of the time","4 A little of the time","5 None of the time")
levels(data$rand36_33.factor)=c("1 Definitely true","2 Mostly true","3 Dont know","4 Mostly false","5 Definitely false")
levels(data$rand36_34.factor)=c("1 Definitely true","2 Mostly true","3 Dont know","4 Mostly false","5 Definitely false")
levels(data$rand36_35.factor)=c("1 Definitely true","2 Mostly true","3 Dont know","4 Mostly false","5 Definitely false")
levels(data$rand36_36.factor)=c("1 Definitely true","2 Mostly true","3 Dont know","4 Mostly false","5 Definitely false")
levels(data$rand_36_item_sf_health_survey_instrument_version_1_complete.factor)=c("Incomplete","Unverified","Complete")

####
#
# Data cleaning. 
# remove duplicates, non-consenting entries and identifiers
#
####

#filter data to remove non-Australians and no consent to participate
#data.clean2 <- subset(data.clean, residence != 0 & sex_diff_bool != 0 & consentb)

data.base <- subset(data, residence != 0 & !is.na(residence) & consentb != 0  & !is.na(consentb))

#check for the duplications that had two entries with the same number of completed sections, then take most recent
multi.entry <- data.base %>%
  group_by(emailid) %>%
  filter(n()>1) %>%
  summarise(entries=n())
# # 120 participants suplicated their entries
sum(multi.entry$entries) - 120
# a total of 147 entries need to be removed to get single entries

multi.rm <- data.base %>%
  select(record_id,
         emailid,
         intro_complete,
         consent_form_complete,
         demographics_complete,
         disease_and_diagnosis_complete,
         symptom_management_complete,
         mental_health_complete,
         neuroqol_bank_v10_fatigue_complete,
         alcohol_dependence_scale_complete,
         rand_36_item_sf_health_survey_instrument_version_1_complete)%>%
  mutate(count_consent = apply(.[,3:11], 1, count_comp)) %>%
  group_by(emailid) %>%
  slice(which.max(count_consent))
# #Problem: takes first instance of which.max encountering TRUE, so the first entry is taken if duplicated full submissions though correct number of end entries

multi.rm.2 <- data.base %>% 
  select(record_id, 
         emailid,
         intro_complete,
         consent_form_complete,
         demographics_complete,
         disease_and_diagnosis_complete,
         symptom_management_complete,
         mental_health_complete,
         neuroqol_bank_v10_fatigue_complete,
         alcohol_dependence_scale_complete,
         rand_36_item_sf_health_survey_instrument_version_1_complete)%>%
  mutate(count_consent = apply(.[,3:11], 1, count_comp)) %>%
  group_by(emailid) %>% 
  arrange(desc(record_id)) %>%
  slice(which.max(count_consent)) %>%
  ungroup()
#make tibble with email and record entries to see differences between the two which.max approaches
take.test <- tibble(
  bol = multi.rm$record_id,
  bol.em = multi.rm$emailid,
  bol2 = multi.rm.2$record_id,
  bol2.em = multi.rm.2$emailid
)
# #check differences in approach
query <- take.test %>%
  filter(bol != bol2)
# #approach 2 worked!
# #time to re-apply this to the initial data.base 
data.base <- multi.rm.2 %>%
  select(record_id) %>%
  left_join(data.base, by="record_id")

#df with all the consent, ID and emails in it along with whether they'd like to be contacted down the track
data.identifiable <- data.base %>% select(record_id,
                                         emailid,
                                         sex_diff_bool,
                                         consentb,
                                         consentb_2)
write.table(x = data.identifiable,
           file = "Survey-identifiable-info_final.tsv",
           row.names = FALSE,
           sep = "\t")

# df with record IDs of participants that wish to be contacted
future.contact <- data.base %>%
  filter(consentb == "2") %>%
  select(record_id,
         emailid)
write.table(x = future.contact,
           file = "Survey-participants-consent-future-studies_final.tsv",
           row.names = FALSE,
           sep = "\t")

#df with participants that wish to be informed of survey results
results.contact <- data.base %>%
  filter(consentb_2 == "1") %>%
  select(record_id,
         emailid)
write.table(x = results.contact,
            file = "Survey-participants-contact-with-results_final.tsv",
            row.names = FALSE,
            sep = "\t")

#df without identifying info, use this for further analysis
data.clean <- subset(data.base, select = -c(emailid, 
                                            sex_diff_bool,  
                                            consentb,
                                            consentb_2))
###############################
### Trouble shooting
###############################

##check why some entries =/= HC or Chron assignment
unique(data.clean$diseasestat) %>% str
disease.na <- data.clean %>%
  filter(is.na(diseasestat))
#only two of the 152 entries have useable data. ID 1797 & 941 later on have dx, so manually change diseasestat to 1 and remove the others
#check that those 150 weren't duplicates with more useful data
na.record <- disease.na %>%
  select(record_id)
na.check <- data %>%
  group_by(emailid) %>%
  filter(any(record_id %in% na.record$record_id)) %>%
  filter(any(duplicated(emailid)))
#10 of the 150 were in duplicates, however all 20 entries are almost entirely NAs with no useful data. So rm 150
data.clean <- data.clean %>%
  mutate_at(vars(c(diseasestat)), ~ ifelse(record_id == "1797" |
                                             record_id == "941" , "1", .))
#double check correct change
data.clean %>% filter(record_id =="941") %>% select(diseasestat)
#re-label factor
data.clean$diseasestat.factor = factor(data.clean$diseasestat,levels=c("1","0"))
levels(data.clean$diseasestat.factor)=c("Yes","No")
#now remove other 150 entries without disease status or diagnostics
data.clean <- data.clean %>% filter(!is.na(diseasestat))

#check entries that ticked chronic illness but no to self dx and formal dx
progress.check <- data.clean %>%
  subset(., diseasestat == "1" &
             self == "2" &
             formal == "2")
# on closer inspection, 120 of the 150 entries above wrote their entire illness history in the otherchronic section so it was still picked up and sorted into the appropriate categories

saveRDS(data.clean, file = "AD-data-clean.rds")

#clean up
# rm(data, data.base, data.identifiable, results.contact, take.test, query, progress.check,  multi.mail, multi.mail.df, removals)

# ad.uni2 <- data.clean %>%  # for future stuff include: filter_all(any_vars(str_detect(., pattern = ""))) %>%
#   select(autoimmune_id_other) %>%
#   distinct(.)
